added provider and changed file structure
