+++ good_morning/lib/main.dart
@@ -1,8 +1,20 @@
 import 'package:flutter/material.dart';
+import 'package:good_morning/utils/daily_fact/daily_fact_list.dart';
 import '/ui/home_page.dart';
+import 'package:provider/provider.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(
+    MultiProvider(
+      providers: [
+        ChangeNotifierProvider(
+          create: (context) => DailyFactList(),
+        ),
+        // other providers here..
+      ],
+      child: const MyApp(),
+    ),
+  );
 }
 
 class MyApp extends StatelessWidget {
