+++ good_morning/lib/ui/home_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:good_morning/ui/daily_fact_ui.dart';
+import 'package:good_morning/ui/daily_fact/daily_fact_ui.dart';
 
 class HomePage extends StatefulWidget {
   @override
