+++ good_morning/pubspec.yaml
@@ -30,6 +30,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  provider: any
 
 
   # The following adds the Cupertino Icons font to your application.
